Refactoring functions project to tidy up
